feat: refactor Breadcrumb component for improved structure and functionality
This update refactors the Breadcrumb component by consolidating the BreadcrumbItem logic into its own component, enhancing maintainability and readability. The rendering logic has been streamlined to support both full and truncated breadcrumb displays, with added dropdown functionality for hidden items. Additionally, utility functions have been documented for better clarity and understanding of their purpose. The BreadcrumbItemProps interface has been extended to include internal properties, ensuring a clear separation of concerns.
