+++ lib/components/breadcrumb/index.ts
@@ -1,2 +1,4 @@
 export { default as Breadcrumb } from './Breadcrumb';
-export * from './types'; 
+export { BreadcrumbItem } from './BreadcrumbItem';
+export { BreadcrumbVariant } from './types';
+export type { BreadcrumbProps, BreadcrumbItemProps } from './types';
